refactor: update include paths for probabilistic occupancy grid map
