+++ perception/probabilistic_occupancy_grid_map/lib/costmap_2d/occupancy_grid_map_projective.cpp
@@ -12,13 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "probabilistic_occupancy_grid_map/pointcloud_based_occupancy_grid_map/occupancy_grid_map_fixed.hpp"
+#include "probabilistic_occupancy_grid_map/pointcloud_based_occupancy_grid_map/occupancy_grid_map_projective.hpp"
 
-#include "probabilistic_occupancy_grid_map/cost_value.hpp"
-#include "probabilistic_occupancy_grid_map/utils/utils.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/cost_value/cost_value.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/utils/utils.hpp"
 
 #include <autoware/universe_utils/math/unit_conversion.hpp>
 #include <grid_map_costmap_2d/grid_map_costmap_2d.hpp>
+#include <grid_map_ros/grid_map_ros.hpp>
 #include <pcl_ros/transforms.hpp>
 
 #include <sensor_msgs/point_cloud2_iterator.hpp>
@@ -37,21 +38,21 @@ namespace costmap_2d
 {
 using sensor_msgs::PointCloud2ConstIterator;
 
-OccupancyGridMapFixedBlindSpot::OccupancyGridMapFixedBlindSpot(
+OccupancyGridMapProjectiveBlindSpot::OccupancyGridMapProjectiveBlindSpot(
   const unsigned int cells_size_x, const unsigned int cells_size_y, const float resolution)
 : OccupancyGridMapInterface(cells_size_x, cells_size_y, resolution)
 {
 }
 
 /**
- * @brief update Gridmap with PointCloud
+ * @brief update Gridmap with PointCloud in 3D manner
  *
  * @param raw_pointcloud raw point cloud on a certain frame (usually base_link)
  * @param obstacle_pointcloud raw point cloud on a certain frame (usually base_link)
  * @param robot_pose frame of the input point cloud (usually base_link)
  * @param scan_origin manually chosen grid map origin frame
  */
-void OccupancyGridMapFixedBlindSpot::updateWithPointCloud(
+void OccupancyGridMapProjectiveBlindSpot::updateWithPointCloud(
   const PointCloud2 & raw_pointcloud, const PointCloud2 & obstacle_pointcloud,
   const Pose & robot_pose, const Pose & scan_origin)
 {
@@ -71,52 +72,81 @@ void OccupancyGridMapFixedBlindSpot::updateWithPointCloud(
   utils::transformPointcloud(map_raw_pointcloud, scan2map_pose, scan_raw_pointcloud);
   utils::transformPointcloud(map_obstacle_pointcloud, scan2map_pose, scan_obstacle_pointcloud);
 
-  // Create angle bins and sort by distance
-  struct BinInfo
+  // Create angle bins and sort points by range
+  struct BinInfo3D
   {
-    BinInfo() = default;
-    BinInfo(const double _range, const double _wx, const double _wy)
-    : range(_range), wx(_wx), wy(_wy)
+    explicit BinInfo3D(
+      const double _range = 0.0, const double _wx = 0.0, const double _wy = 0.0,
+      const double _wz = 0.0, const double _projection_length = 0.0,
+      const double _projected_wx = 0.0, const double _projected_wy = 0.0)
+    : range(_range),
+      wx(_wx),
+      wy(_wy),
+      wz(_wz),
+      projection_length(_projection_length),
+      projected_wx(_projected_wx),
+      projected_wy(_projected_wy)
     {
     }
     double range;
     double wx;
     double wy;
+    double wz;
+    double projection_length;
+    double projected_wx;
+    double projected_wy;
   };
-  std::vector</*angle bin*/ std::vector<BinInfo>> obstacle_pointcloud_angle_bins;
-  std::vector</*angle bin*/ std::vector<BinInfo>> raw_pointcloud_angle_bins;
+
+  std::vector</*angle bin*/ std::vector<BinInfo3D>> obstacle_pointcloud_angle_bins;
+  std::vector</*angle bin*/ std::vector<BinInfo3D>> raw_pointcloud_angle_bins;
   obstacle_pointcloud_angle_bins.resize(angle_bin_size);
   raw_pointcloud_angle_bins.resize(angle_bin_size);
   for (PointCloud2ConstIterator<float> iter_x(scan_raw_pointcloud, "x"),
        iter_y(scan_raw_pointcloud, "y"), iter_wx(map_raw_pointcloud, "x"),
-       iter_wy(map_raw_pointcloud, "y");
-       iter_x != iter_x.end(); ++iter_x, ++iter_y, ++iter_wx, ++iter_wy) {
+       iter_wy(map_raw_pointcloud, "y"), iter_wz(map_raw_pointcloud, "z");
+       iter_x != iter_x.end(); ++iter_x, ++iter_y, ++iter_wx, ++iter_wy, ++iter_wz) {
     const double angle = atan2(*iter_y, *iter_x);
     const int angle_bin_index = (angle - min_angle) / angle_increment;
     raw_pointcloud_angle_bins.at(angle_bin_index)
-      .push_back(BinInfo(std::hypot(*iter_y, *iter_x), *iter_wx, *iter_wy));
+      .emplace_back(std::hypot(*iter_y, *iter_x), *iter_wx, *iter_wy, *iter_wz);
   }
   for (auto & raw_pointcloud_angle_bin : raw_pointcloud_angle_bins) {
     std::sort(raw_pointcloud_angle_bin.begin(), raw_pointcloud_angle_bin.end(), [](auto a, auto b) {
       return a.range < b.range;
     });
   }
-  // create and sort bin for obstacle pointcloud
+  // Create obstacle angle bins and sort points by range
   for (PointCloud2ConstIterator<float> iter_x(scan_obstacle_pointcloud, "x"),
-       iter_y(scan_obstacle_pointcloud, "y"), iter_wx(map_obstacle_pointcloud, "x"),
-       iter_wy(map_obstacle_pointcloud, "y");
-       iter_x != iter_x.end(); ++iter_x, ++iter_y, ++iter_wx, ++iter_wy) {
+       iter_y(scan_obstacle_pointcloud, "y"), iter_z(scan_obstacle_pointcloud, "z"),
+       iter_wx(map_obstacle_pointcloud, "x"), iter_wy(map_obstacle_pointcloud, "y"),
+       iter_wz(map_obstacle_pointcloud, "z");
+       iter_x != iter_x.end(); ++iter_x, ++iter_y, ++iter_z, ++iter_wx, ++iter_wy, ++iter_wz) {
+    const double scan_z = scan_origin.position.z - robot_pose.position.z;
+    const double obstacle_z = (*iter_wz) - robot_pose.position.z;
+    const double dz = scan_z - obstacle_z;
     const double angle = atan2(*iter_y, *iter_x);
     const int angle_bin_index = (angle - min_angle) / angle_increment;
-    const double range = std::hypot(*iter_y, *iter_x);
+    const double range = std::hypot(*iter_x, *iter_y);
     // Ignore obstacle points exceed the range of the raw points
     if (raw_pointcloud_angle_bins.at(angle_bin_index).empty()) {
       continue;  // No raw point in this angle bin
     } else if (range > raw_pointcloud_angle_bins.at(angle_bin_index).back().range) {
       continue;  // Obstacle point exceeds the range of the raw points
     }
-    obstacle_pointcloud_angle_bins.at(angle_bin_index)
-      .push_back(BinInfo(range, *iter_wx, *iter_wy));
+    if (dz > projection_dz_threshold_) {
+      const double ratio = obstacle_z / dz;
+      const double projection_length = range * ratio;
+      const double projected_wx = (*iter_wx) + ((*iter_wx) - scan_origin.position.x) * ratio;
+      const double projected_wy = (*iter_wy) + ((*iter_wy) - scan_origin.position.y) * ratio;
+      obstacle_pointcloud_angle_bins.at(angle_bin_index)
+        .emplace_back(
+          range, *iter_wx, *iter_wy, *iter_wz, projection_length, projected_wx, projected_wy);
+    } else {
+      obstacle_pointcloud_angle_bins.at(angle_bin_index)
+        .emplace_back(
+          range, *iter_wx, *iter_wy, *iter_wz, std::numeric_limits<double>::infinity(),
+          std::numeric_limits<double>::infinity(), std::numeric_limits<double>::infinity());
+    }
   }
   for (auto & obstacle_pointcloud_angle_bin : obstacle_pointcloud_angle_bins) {
     std::sort(
@@ -124,60 +154,87 @@ void OccupancyGridMapFixedBlindSpot::updateWithPointCloud(
       [](auto a, auto b) { return a.range < b.range; });
   }
 
+  grid_map::Costmap2DConverter<grid_map::GridMap> converter;
+  if (pub_debug_grid_) {
+    debug_grid_.clearAll();
+    debug_grid_.setFrameId("map");
+    debug_grid_.setGeometry(
+      grid_map::Length(size_x_ * resolution_, size_y_ * resolution_), resolution_,
+      grid_map::Position(
+        origin_x_ + size_x_ * resolution_ / 2.0, origin_y_ + size_y_ * resolution_ / 2.0));
+  }
+
+  auto is_visible_beyond_obstacle = [&](const BinInfo3D & obstacle, const BinInfo3D & raw) -> bool {
+    if (raw.range < obstacle.range) {
+      return false;
+    }
+
+    if (std::isinf(obstacle.projection_length)) {
+      return false;
+    }
+
+    // y = ax + b
+    const double a = -(scan_origin.position.z - robot_pose.position.z) /
+                     (obstacle.range + obstacle.projection_length);
+    const double b = scan_origin.position.z;
+    return raw.wz > (a * raw.range + b);
+  };
+
   // First step: Initialize cells to the final point with freespace
   for (size_t bin_index = 0; bin_index < obstacle_pointcloud_angle_bins.size(); ++bin_index) {
-    auto & raw_pointcloud_angle_bin = raw_pointcloud_angle_bins.at(bin_index);
+    const auto & raw_pointcloud_angle_bin = raw_pointcloud_angle_bins.at(bin_index);
 
-    BinInfo end_distance;
+    BinInfo3D ray_end;
     if (raw_pointcloud_angle_bin.empty()) {
       continue;
     } else {
-      end_distance = raw_pointcloud_angle_bin.back();
+      ray_end = raw_pointcloud_angle_bin.back();
     }
     raytrace(
-      scan_origin.position.x, scan_origin.position.y, end_distance.wx, end_distance.wy,
-      occupancy_cost_value::FREE_SPACE);
+      scan_origin.position.x, scan_origin.position.y, ray_end.wx, ray_end.wy,
+      cost_value::FREE_SPACE);
   }
+
+  if (pub_debug_grid_)
+    converter.addLayerFromCostmap2D(*this, "filled_free_to_farthest", debug_grid_);
 
   // Second step: Add unknown cell
   for (size_t bin_index = 0; bin_index < obstacle_pointcloud_angle_bins.size(); ++bin_index) {
-    auto & obstacle_pointcloud_angle_bin = obstacle_pointcloud_angle_bins.at(bin_index);
-    auto & raw_pointcloud_angle_bin = raw_pointcloud_angle_bins.at(bin_index);
+    const auto & obstacle_pointcloud_angle_bin = obstacle_pointcloud_angle_bins.at(bin_index);
+    const auto & raw_pointcloud_angle_bin = raw_pointcloud_angle_bins.at(bin_index);
     auto raw_distance_iter = raw_pointcloud_angle_bin.begin();
     for (size_t dist_index = 0; dist_index < obstacle_pointcloud_angle_bin.size(); ++dist_index) {
       // Calculate next raw point from obstacle point
+      const auto & obstacle_bin = obstacle_pointcloud_angle_bin.at(dist_index);
       while (raw_distance_iter != raw_pointcloud_angle_bin.end()) {
-        if (
-          raw_distance_iter->range <
-          obstacle_pointcloud_angle_bin.at(dist_index).range + distance_margin_)
+        if (!is_visible_beyond_obstacle(obstacle_bin, *raw_distance_iter))
           raw_distance_iter++;
         else
           break;
       }
 
-      // There is no point far than the obstacle point.
-      const bool no_freespace_point = (raw_distance_iter == raw_pointcloud_angle_bin.end());
+      // There is no point farther than the obstacle point.
+      const bool no_visible_point_beyond = (raw_distance_iter == raw_pointcloud_angle_bin.end());
+      if (no_visible_point_beyond) {
+        const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
+        raytrace(
+          source.wx, source.wy, source.projected_wx, source.projected_wy,
+          cost_value::NO_INFORMATION);
+        break;
+      }
 
       if (dist_index + 1 == obstacle_pointcloud_angle_bin.size()) {
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
-        if (!no_freespace_point) {
-          const auto & target = *raw_distance_iter;
-          raytrace(
-            source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::NO_INFORMATION);
-          setCellValue(target.wx, target.wy, occupancy_cost_value::FREE_SPACE);
-        }
+        raytrace(
+          source.wx, source.wy, source.projected_wx, source.projected_wy,
+          cost_value::NO_INFORMATION);
         continue;
       }
 
       auto next_obstacle_point_distance = std::abs(
         obstacle_pointcloud_angle_bin.at(dist_index + 1).range -
         obstacle_pointcloud_angle_bin.at(dist_index).range);
-      if (next_obstacle_point_distance <= distance_margin_) {
-        continue;
-      } else if (no_freespace_point) {
-        const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
-        const auto & target = obstacle_pointcloud_angle_bin.at(dist_index + 1);
-        raytrace(source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::NO_INFORMATION);
+      if (next_obstacle_point_distance <= obstacle_separation_threshold_) {
         continue;
       }
 
@@ -186,24 +243,26 @@ void OccupancyGridMapFixedBlindSpot::updateWithPointCloud(
       if (next_raw_distance < next_obstacle_point_distance) {
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         const auto & target = *raw_distance_iter;
-        raytrace(source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::NO_INFORMATION);
-        setCellValue(target.wx, target.wy, occupancy_cost_value::FREE_SPACE);
+        raytrace(source.wx, source.wy, target.wx, target.wy, cost_value::NO_INFORMATION);
+        setCellValue(target.wx, target.wy, cost_value::FREE_SPACE);
         continue;
       } else {
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         const auto & target = obstacle_pointcloud_angle_bin.at(dist_index + 1);
-        raytrace(source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::NO_INFORMATION);
+        raytrace(source.wx, source.wy, target.wx, target.wy, cost_value::NO_INFORMATION);
         continue;
       }
     }
   }
+
+  if (pub_debug_grid_) converter.addLayerFromCostmap2D(*this, "added_unknown", debug_grid_);
 
   // Third step: Overwrite occupied cell
   for (size_t bin_index = 0; bin_index < obstacle_pointcloud_angle_bins.size(); ++bin_index) {
     auto & obstacle_pointcloud_angle_bin = obstacle_pointcloud_angle_bins.at(bin_index);
     for (size_t dist_index = 0; dist_index < obstacle_pointcloud_angle_bin.size(); ++dist_index) {
       const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
-      setCellValue(source.wx, source.wy, occupancy_cost_value::LETHAL_OBSTACLE);
+      setCellValue(source.wx, source.wy, cost_value::LETHAL_OBSTACLE);
 
       if (dist_index + 1 == obstacle_pointcloud_angle_bin.size()) {
         continue;
@@ -212,20 +271,31 @@ void OccupancyGridMapFixedBlindSpot::updateWithPointCloud(
       auto next_obstacle_point_distance = std::abs(
         obstacle_pointcloud_angle_bin.at(dist_index + 1).range -
         obstacle_pointcloud_angle_bin.at(dist_index).range);
-      if (next_obstacle_point_distance <= distance_margin_) {
+      if (next_obstacle_point_distance <= obstacle_separation_threshold_) {
         const auto & source = obstacle_pointcloud_angle_bin.at(dist_index);
         const auto & target = obstacle_pointcloud_angle_bin.at(dist_index + 1);
-        raytrace(source.wx, source.wy, target.wx, target.wy, occupancy_cost_value::LETHAL_OBSTACLE);
+        raytrace(source.wx, source.wy, target.wx, target.wy, cost_value::LETHAL_OBSTACLE);
         continue;
       }
     }
   }
+
+  if (pub_debug_grid_) converter.addLayerFromCostmap2D(*this, "added_obstacle", debug_grid_);
+  if (pub_debug_grid_) {
+    debug_grid_map_publisher_ptr_->publish(grid_map::GridMapRosConverter::toMessage(debug_grid_));
+  }
 }
 
-void OccupancyGridMapFixedBlindSpot::initRosParam(rclcpp::Node & node)
+void OccupancyGridMapProjectiveBlindSpot::initRosParam(rclcpp::Node & node)
 {
-  distance_margin_ =
-    node.declare_parameter<double>("OccupancyGridMapFixedBlindSpot.distance_margin");
+  projection_dz_threshold_ =
+    node.declare_parameter<double>("OccupancyGridMapProjectiveBlindSpot.projection_dz_threshold");
+  obstacle_separation_threshold_ = node.declare_parameter<double>(
+    "OccupancyGridMapProjectiveBlindSpot.obstacle_separation_threshold");
+  pub_debug_grid_ =
+    node.declare_parameter<bool>("OccupancyGridMapProjectiveBlindSpot.pub_debug_grid");
+  debug_grid_map_publisher_ptr_ = node.create_publisher<grid_map_msgs::msg::GridMap>(
+    "~/debug/grid_map", rclcpp::QoS(1).durability_volatile());
 }
 
 }  // namespace costmap_2d
