+++ perception/probabilistic_occupancy_grid_map/lib/updater/occupancy_grid_map_log_odds_bayes_filter_updater.cpp
@@ -14,7 +14,7 @@
 
 #include "probabilistic_occupancy_grid_map/updater/occupancy_grid_map_log_odds_bayes_filter_updater.hpp"
 
-#include "probabilistic_occupancy_grid_map/cost_value.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/cost_value/cost_value.hpp"
 
 #include <algorithm>
 
@@ -35,7 +35,7 @@ inline unsigned char OccupancyGridMapLOBFUpdater::applyLOBF(
   using fusion_policy::convertProbabilityToChar;
   using fusion_policy::log_odds_fusion::logOddsFusion;
 
-  constexpr unsigned char unknown = occupancy_cost_value::NO_INFORMATION;
+  constexpr unsigned char unknown = cost_value::NO_INFORMATION;
   constexpr unsigned char unknown_margin = 1;
   /* Tau and ST decides how fast the observation decay to the unknown status*/
   constexpr double tau = 0.75;
