+++ perception/probabilistic_occupancy_grid_map/test/fusion_policy_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "probabilistic_occupancy_grid_map/cost_value.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/cost_value/cost_value.hpp"
 #include "probabilistic_occupancy_grid_map/fusion/single_frame_fusion_policy.hpp"
 
 #include <gtest/gtest.h>
